requirements.txt
# LLM Integration
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0

# Data Processing
datasets>=2.16.0
pandas>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0

# Visualization & Analysis
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
jupyter>=1.0.0
pyyaml>=6.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
